+++ fla/ops/simple_gla/fused_recurrent.py
@@ -30,7 +30,7 @@ def fused_recurrent_simple_gla(
         g (torch.Tensor):
             Forget gates of shape `[B, T, H]`.
             Compared to GLA, the gating is head-wise instead of elementwise.
-        scale (Optional[int]):
+        scale (Optional[float]):
             Scale factor for the attention scores.
             If not provided, it will default to `1 / sqrt(K)`. Default: `None`.
         initial_state (Optional[torch.Tensor]):
